reporte

+++ src/Celulares.java
@@ -5,7 +5,7 @@ public class Celulares {
 	public static double almacenamiento0 = 256;
 	public static double bateria0 = 5000;
 	public static double camara0 = 200;
-	public static double cant0 = 0;
+	public static int cant0 = 0;
 	
 	
 
@@ -14,7 +14,7 @@ public class Celulares {
 	public static double almacenamiento1 = 256;
 	public static double bateria1 = 5000;
 	public static double camara1 = 50;
-	public static double cant1 = 0;
+	public static int cant1 = 0;
 	
 
 	public static String modelo2 = "POCO X8 Pro";
@@ -22,7 +22,7 @@ public class Celulares {
 	public static double almacenamiento2 = 512;
 	public static double bateria2 = 5500;
 	public static double camara2 = 64;
-	public static double cant2 = 0;
+	public static int cant2 = 0;
 	
 
 	public static String modelo3 = "Google Pixel 10a";
@@ -30,7 +30,7 @@ public class Celulares {
 	public static double almacenamiento3 = 128;
 	public static double bateria3 = 4700;
 	public static double camara3 = 50;
-	public static double cant3 = 0;
+	public static int cant3 = 0;
 	
 
 	public static String modelo4 = "Honor Magic7 Lite";
@@ -38,14 +38,14 @@ public class Celulares {
 	public static double almacenamiento4 = 256;
 	public static double bateria4 = 6600;
 	public static double camara4 = 108;
-	public static double cant4 = 0;
+	public static int cant4 = 0;
 
 	public static String modelo5 = "Vivo V70 FE";
 	public static double precio5 = 1899;
 	public static double almacenamiento5 = 256;
 	public static double bateria5 = 5250;
 	public static double camara5 = 200;
-	public static double cant5 = 0;
+	public static int cant5 = 0;
 	
 	
 	public static double porcentaje1 = 7.5;
